fix: show icon and proper name for lightning node backend on about page
