+++ frontend/src/screens/setup/SetupNode.tsx
@@ -1,60 +1,20 @@
-import React, { ReactElement } from "react";
+import React from "react";
 import { useNavigate } from "react-router";
 import Container from "src/components/Container";
 import TwoColumnLayoutHeader from "src/components/TwoColumnLayoutHeader";
-import { LDKIcon } from "src/components/icons/LDK";
-import { PhoenixdIcon } from "src/components/icons/Phoenixd";
 import { Button } from "src/components/ui/button";
 import { cn } from "src/lib/utils";
 import { BackendType } from "src/types";
 
-import bark from "src/assets/images/node/bark.jpg";
-import cashu from "src/assets/images/node/cashu.png";
-import cln from "src/assets/images/node/cln.png";
-import lnd from "src/assets/images/node/lnd.png";
 import { backendTypeConfigs } from "src/lib/backendType";
 import useSetupStore from "src/state/SetupStore";
 
-type BackendTypeDisplayConfig = {
-  title: string;
-  icon: ReactElement;
-};
-
-const backendTypeDisplayConfigs: Partial<
-  Record<BackendType, BackendTypeDisplayConfig>
-> = {
-  LDK: {
-    title: "LDK",
-    icon: <LDKIcon />,
-  },
-  PHOENIX: {
-    title: "phoenixd",
-    icon: <PhoenixdIcon />,
-  },
-  LND: {
-    title: "LND",
-    icon: <img src={lnd} />,
-  },
-  CASHU: {
-    title: "Cashu Mint",
-    icon: <img src={cashu} />,
-  },
-  CLN: {
-    title: "CLN",
-    icon: <img src={cln} />,
-  },
-  BARK: {
-    title: "Bark",
-    icon: <img src={bark} />,
-  },
-};
-
-const backendTypeDisplayConfigList = Object.entries(
-  backendTypeDisplayConfigs
-).map((entry) => ({
-  ...entry[1],
-  backendType: entry[0] as BackendType,
-}));
+const backendTypeDisplayConfigList = Object.entries(backendTypeConfigs).map(
+  (entry) => ({
+    ...entry[1],
+    backendType: entry[0] as BackendType,
+  })
+);
 
 export function SetupNode() {
   const navigate = useNavigate();
